+++ packages/core/src/languages/__tests__/builtin.test.ts
@@ -6,6 +6,9 @@ import HtmlAttributeMangler from "../../manglers/html-attributes";
 import HtmlIdMangler from "../../manglers/html-ids";
 
 import BuiltInLanguageSupport from "../builtin";
+import CssLanguageSupport from "../css";
+import HtmlLanguageSupport from "../html";
+import JavaScriptLanguageSupport from "../javascript";
 
 suite("Built-in Language Supports", function() {
   test(`has support for the ${CssClassMangler._ID} mangler`, function() {
@@ -34,5 +37,23 @@ suite("Built-in Language Supports", function() {
 
     const expressions = plugin.getExpressionsFor(HtmlIdMangler._ID);
     expect(expressions).to.have.length.above(0);
+  });
+
+  test("get languages", function() {
+    const plugin = new BuiltInLanguageSupport();
+
+    const result = plugin.getLanguages();
+
+    for (const language of new CssLanguageSupport().getLanguages()) {
+      expect(result).to.include(language);
+    }
+
+    for (const language of new HtmlLanguageSupport().getLanguages()) {
+      expect(result).to.include(language);
+    }
+
+    for (const language of new JavaScriptLanguageSupport().getLanguages()) {
+      expect(result).to.include(language);
+    }
   });
 });