+++ packages/core/src/languages/__tests__/css.test.ts
@@ -35,4 +35,11 @@ suite("Built-in CSS Language Support", function() {
     const expressions = plugin.getExpressionsFor(HtmlIdMangler._ID);
     expect(expressions).to.have.length.above(0);
   });
+
+  test("get languages", function() {
+    const plugin = new CssLanguageSupport();
+
+    const result = plugin.getLanguages();
+    expect(result).to.include("css");
+  });
 });